requirements.txt
Django==2.1.5
djangorestframework==3.9.0
pytz==2018.9
